Install react-navigation
