+++ IntentionalWalkApp/package.json
@@ -10,11 +10,16 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.6",
     "react": "16.9.0",
     "react-native": "0.61.5",
     "react-native-gesture-handler": "^1.5.3",
     "react-native-progress": "^4.0.3",
-    "react-navigation": "^4.1.0"
+    "react-native-reanimated": "^1.7.0",
+    "react-native-safe-area-context": "^0.6.4",
+    "react-native-screens": "^2.0.0-alpha.32",
+    "react-navigation": "^4.1.0",
+    "react-navigation-stack": "^2.0.16"
   },
   "devDependencies": {
     "@babel/core": "^7.6.2",
