+++ IntentionalWalkApp/screens/welcome.js
@@ -1,0 +1,10 @@
+import React, {useState} from 'react';
+import {StyleSheet, View, Text, TouchableOpacity} from 'react-native';
+
+export default function Welcome({navigation}) {
+  return (
+    <View>
+      <Text>Welcome</Text>
+    </View>
+  );
+}
